Add `secure` middleware, with sensible values
